Only rebuild XEPs that have depenencies changed
repository.AllXEPs will check if any of the depencenies for the
to-be-generated output is newer than existing output. If it is, it
will rebuild the PDF and HTML, if not, it will skip the XEP.

Test-Information:

Test platform Debian 8 x64
Building XEPs from scratch works.
Breaking off the build and restarting doesn't rebuild already
up-to-date output.
Touching xep.ent causes all XEPs to rebuild.